Refactor Data processing
