+++ apps/charterafrica/src/lib/ecosystem/ecosystem.js
@@ -1,4 +1,9 @@
-import getListFromAirtable from "@/charterafrica/lib/ecosystem/airtable";
+import {
+  getListFromAirtable,
+  processToolFromAirtable,
+  processOrganisationFromAirTable,
+  processContributorFromAirtable,
+} from "@/charterafrica/lib/ecosystem/airtable";
 import {
   createCollection,
   ORGANIZATION_COLLECTION,
@@ -16,117 +21,6 @@ export const updateEcosystemList = async (req, res) => {
   };
   const response = await getListFromAirtable(args);
   res.status(200).json(response);
-};
-
-const processOrganisationFromAirTable = async (data) => {
-  const config = await api.findGlobal(ECOSYSTEM_CONFIG, {});
-  const description = {
-    en: data[config.organisationDescription.english],
-    pt: data[config.organisationDescription.portuguese],
-    fr: data[config.organisationDescription.french],
-  };
-  const unLocalizedData = {
-    airtableId: data.id,
-    externalId: data[config.organisationUserName] || data.id,
-    name: data[config.organisationName]?.[0],
-    type: data[config.organisationType],
-    repoLink: data[config.organisationRepoLink],
-    donors: [], // data.Donors, UPDATE when source is sanitized
-    partners: [], // data.Partners,
-  };
-  return {
-    en: {
-      ...unLocalizedData,
-      description: description.en,
-    },
-    fr: {
-      ...unLocalizedData,
-      description: description.fr,
-    },
-    pt: {
-      ...unLocalizedData,
-      description: description.pt,
-    },
-  };
-};
-
-const processContributorFromAirtable = async (data) => {
-  const config = await api.findGlobal(ECOSYSTEM_CONFIG, {});
-  const description = {
-    en: data[config.contributorDescription.english],
-    pt: data[config.contributorDescription.portuguese],
-    fr: data[config.contributorDescription.french],
-  };
-  const defaultData = {
-    airtableId: data.id,
-    externalId: data[config.contributorUserName] || data.id,
-    repoLink: `https://github.com/${data.id}`,
-    donors: [], // data.Donors,
-    partners: [], // data.Partners,
-  };
-  return {
-    en: {
-      ...defaultData,
-      description: description.en,
-    },
-    fr: {
-      ...defaultData,
-      description: description.fr,
-    },
-    pt: {
-      ...defaultData,
-      description: description.pt,
-    },
-  };
-};
-
-const processToolFromAirtable = async (data) => {
-  const config = await api.findGlobal(ECOSYSTEM_CONFIG, {});
-  const theme = {
-    en: data[config.toolTheme.english]?.[0],
-    pt: data[config.toolTheme.portuguese]?.[0],
-    fr: data[config.toolTheme.french]?.[0],
-  };
-  const description = {
-    en: data[config.toolsDescription.english] || "",
-    pt: data[config.toolsDescription.portuguese] || "",
-    fr: data[config.toolsDescription.french] || "",
-  };
-
-  const operatingCountries = [];
-  const homeCountry = data[config.homeCountry];
-  const defaultData = {
-    airtableId: data.id,
-    externalId: data[config.toolExternalId],
-    repoLink: data[config.toolRepoLink],
-    name: data[config.toolName],
-    link: data[config.toolLink],
-    operatingCountries,
-    contributors: data[config.toolContributors],
-    organisation: data[config.toolOrganisation]?.[0],
-    donors: [], // data.Donors,
-    partners: [], //  data.Partners,
-    homeCountry,
-    otherSocialMedia: [],
-  };
-
-  return {
-    en: {
-      ...defaultData,
-      description: description.en,
-      theme: theme.en,
-    },
-    fr: {
-      ...defaultData,
-      description: description.fr,
-      theme: theme.fr,
-    },
-    pt: {
-      ...defaultData,
-      description: description.pt,
-      theme: theme.pt,
-    },
-  };
 };
 
 export const updateEcosystemContent = async (req, res) => {
